Finished implementing area iteration and standard deviation calculation
Cleaned old functions in comp_area.c/h which weren't needed and the values pointed to constants instead of the
calculations now stored in the engine_t strcut

Wrote some specifications in the TODO file to have a clearer undestanding of where are some of the needed values
for the future implementations planned

Tested compilation 15 of August
Platform: Windows 10 with Visual Studio Community 2022

+++ include/TREL.h
@@ -6,7 +6,7 @@
 // Types
 // Comp area calculations
 typedef struct {
-    float avg_long_area, avg_trans_area, avg_burn_area, burn_std_deviation, burn_sum_diff;
+    double avg_long_area, avg_trans_area, avg_burn_area, burn_std_deviation, burn_sum_diff;
 } comp_area_t;
 
 // Propellent grains structure
